appium_tests/requirements.txt
Appium-Python-Client==5.0.0
attrs==25.3.0
certifi==2025.1.31
colorama==0.4.6
exceptiongroup==1.2.2
frida==16.7.13
frida-tools==13.7.1
h11==0.14.0
idna==3.10
iniconfig==2.1.0
outcome==1.3.0.post0
packaging==25.0
pluggy==1.5.0
prompt_toolkit==3.0.51
Pygments==2.19.1
PySocks==1.7.1
pytest==8.3.5
selenium==4.31.0
sniffio==1.3.1
sortedcontainers==2.4.0
tomli==2.2.1
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==13.1
wsproto==1.2.0
